requirements.txt
# Core
Django>=6.0
djangorestframework>=3.16.1
djangorestframework-simplejwt>=5.5.1
django-cors-headers>=4.9.0
python-dotenv>=0.9.9

# Database
psycopg[binary]>=3.1.0

# Production
gunicorn>=21.2.0

# Development (opcional, para desarrollo local)
python-dotenv>=0.19.0

# Testing (opcional, para testing)
pytest-django>=4.5.2
pytest-cov>=4.0.0

# Code quality (opcional)
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0

# Debug (opcional, solo desarrollo)
django-debug-toolbar>=4.0.0
django-extensions>=3.2.0
